clean up common test methods
