+++ tests/integration/profiling/rule_based_profilers/test_profiler_user_workflows.py
@@ -1,5 +1,4 @@
 import datetime
-import uuid
 from numbers import Number
 from typing import Any, Callable, Dict, List, Optional, Tuple, cast
 from unittest import mock
@@ -12,7 +11,6 @@ from packaging import version
 from ruamel.yaml import YAML
 from ruamel.yaml.comments import CommentedMap
 
-import great_expectations.exceptions as ge_exceptions
 from great_expectations import DataContext
 from great_expectations.core import ExpectationSuite, ExpectationValidationResult
 from great_expectations.core.batch import BatchRequest
@@ -35,6 +33,7 @@ from tests.core.usage_statistics.util import (
 )
 from tests.rule_based_profiler.conftest import ATOL, RTOL
 from tests.rule_based_profiler.parameter_builder.conftest import RANDOM_SEED
+from tests.test_utils import get_validator_with_temporary_expectation_suite
 
 yaml = YAML()
 
@@ -673,30 +672,16 @@ def test_bobby_expect_column_values_to_be_between_auto_yes_default_profiler_conf
 ):
     context: DataContext = bobby_columnar_table_multi_batch_deterministic_data_context
 
-    suite: ExpectationSuite
-
-    expectation_suite_name: str = f"tmp.profiler_suite_{str(uuid.uuid4())[:8]}"
-    try:
-        # noinspection PyUnusedLocal
-        suite = context.get_expectation_suite(
-            expectation_suite_name=expectation_suite_name
-        )
-    except ge_exceptions.DataContextError:
-        suite = context.create_expectation_suite(
-            expectation_suite_name=expectation_suite_name
-        )
-        print(f'Created ExpectationSuite "{suite.expectation_suite_name}".')
-
-    # Use all batches, loaded by Validator, for estimating Expectation argument values.
     batch_request: dict = {
         "datasource_name": "taxi_pandas",
         "data_connector_name": "monthly",
         "data_asset_name": "my_reports",
     }
 
-    validator: Validator = context.get_validator(
-        batch_request=BatchRequest(**batch_request),
-        expectation_suite_name=expectation_suite_name,
+    validator: Validator = get_validator_with_temporary_expectation_suite(
+        component_name="profiler",
+        batch_request=batch_request,
+        data_context=context,
     )
     assert len(validator.batches) == 3
 
@@ -815,20 +800,6 @@ def test_bobby_expect_column_values_to_be_between_auto_yes_default_profiler_conf
 ):
     context: DataContext = bobby_columnar_table_multi_batch_deterministic_data_context
 
-    suite: ExpectationSuite
-
-    expectation_suite_name: str = f"tmp.profiler_suite_{str(uuid.uuid4())[:8]}"
-    try:
-        # noinspection PyUnusedLocal
-        suite = context.get_expectation_suite(
-            expectation_suite_name=expectation_suite_name
-        )
-    except ge_exceptions.DataContextError:
-        suite = context.create_expectation_suite(
-            expectation_suite_name=expectation_suite_name
-        )
-        print(f'Created ExpectationSuite "{suite.expectation_suite_name}".')
-
     batch_request: dict
 
     validator: Validator
@@ -844,9 +815,10 @@ def test_bobby_expect_column_values_to_be_between_auto_yes_default_profiler_conf
         "data_asset_name": "my_reports",
     }
 
-    validator = context.get_validator(
-        batch_request=BatchRequest(**batch_request),
-        expectation_suite_name=expectation_suite_name,
+    validator = get_validator_with_temporary_expectation_suite(
+        component_name="profiler",
+        batch_request=batch_request,
+        data_context=context,
     )
     assert len(validator.batches) == 3
 
@@ -973,6 +945,13 @@ def test_bobby_expect_column_values_to_be_between_auto_yes_default_profiler_conf
         },
     }
 
+    validator = get_validator_with_temporary_expectation_suite(
+        component_name="profiler",
+        batch_request=batch_request,
+        data_context=context,
+    )
+    assert len(validator.batches) == 1
+
     custom_profiler_config = RuleBasedProfilerConfig(
         name="expect_column_values_to_be_between",  # Convention: use "expectation_type" as profiler name.
         config_version=1.0,
@@ -1008,12 +987,6 @@ def test_bobby_expect_column_values_to_be_between_auto_yes_default_profiler_conf
             },
         },
     )
-
-    validator = context.get_validator(
-        batch_request=BatchRequest(**batch_request),
-        expectation_suite_name=expectation_suite_name,
-    )
-    assert len(validator.batches) == 1
 
     result = validator.expect_column_values_to_be_between(
         column=column_name,
@@ -1078,20 +1051,6 @@ def test_bobby_expect_column_values_to_be_between_auto_yes_default_profiler_conf
 
     context: DataContext = bobby_columnar_table_multi_batch_deterministic_data_context
 
-    suite: ExpectationSuite
-
-    expectation_suite_name: str = f"tmp.profiler_suite_{str(uuid.uuid4())[:8]}"
-    try:
-        # noinspection PyUnusedLocal
-        suite = context.get_expectation_suite(
-            expectation_suite_name=expectation_suite_name
-        )
-    except ge_exceptions.DataContextError:
-        suite = context.create_expectation_suite(
-            expectation_suite_name=expectation_suite_name
-        )
-        print(f'Created ExpectationSuite "{suite.expectation_suite_name}".')
-
     batch_request: dict
 
     validator: Validator
@@ -1152,9 +1111,10 @@ def test_bobby_expect_column_values_to_be_between_auto_yes_default_profiler_conf
         "data_asset_name": "my_reports",
     }
 
-    validator = context.get_validator(
-        batch_request=BatchRequest(**batch_request),
-        expectation_suite_name=expectation_suite_name,
+    validator = get_validator_with_temporary_expectation_suite(
+        component_name="profiler",
+        batch_request=batch_request,
+        data_context=context,
     )
     assert len(validator.batches) == 3
 
@@ -1244,9 +1204,10 @@ def test_bobby_expect_column_values_to_be_between_auto_yes_default_profiler_conf
         },
     }
 
-    validator = context.get_validator(
-        batch_request=BatchRequest(**batch_request),
-        expectation_suite_name=expectation_suite_name,
+    validator = get_validator_with_temporary_expectation_suite(
+        component_name="profiler",
+        batch_request=batch_request,
+        data_context=context,
     )
     assert len(validator.batches) == 1
 
@@ -1483,22 +1444,6 @@ def test_bobster_expect_table_row_count_to_be_between_auto_yes_default_profiler_
         bobster_columnar_table_multi_batch_normal_mean_5000_stdev_1000_data_context
     )
 
-    result: ExpectationValidationResult
-
-    suite: ExpectationSuite
-
-    expectation_suite_name: str = f"tmp.profiler_suite_{str(uuid.uuid4())[:8]}"
-    try:
-        # noinspection PyUnusedLocal
-        suite = context.get_expectation_suite(
-            expectation_suite_name=expectation_suite_name
-        )
-    except ge_exceptions.DataContextError:
-        suite = context.create_expectation_suite(
-            expectation_suite_name=expectation_suite_name
-        )
-        print(f'Created ExpectationSuite "{suite.expectation_suite_name}".')
-
     # Use all batches, loaded by Validator, for estimating Expectation argument values.
     batch_request: dict = {
         "datasource_name": "taxi_pandas",
@@ -1506,16 +1451,19 @@ def test_bobster_expect_table_row_count_to_be_between_auto_yes_default_profiler_
         "data_asset_name": "my_reports",
     }
 
-    validator: Validator = context.get_validator(
-        batch_request=BatchRequest(**batch_request),
-        expectation_suite_name=expectation_suite_name,
+    validator: Validator = get_validator_with_temporary_expectation_suite(
+        component_name="profiler",
+        batch_request=batch_request,
+        data_context=context,
     )
     assert len(validator.batches) == 36
 
-    result = validator.expect_table_row_count_to_be_between(
-        result_format="SUMMARY",
-        include_config=True,
-        auto=True,
+    result: ExpectationValidationResult = (
+        validator.expect_table_row_count_to_be_between(
+            result_format="SUMMARY",
+            include_config=True,
+            auto=True,
+        )
     )
     assert result.expectation_config.kwargs["auto"]
 
@@ -1686,29 +1634,16 @@ def test_quentin_expect_column_quantile_values_to_be_between_auto_yes_default_pr
 
     custom_profiler_config: RuleBasedProfilerConfig
 
-    suite: ExpectationSuite
-
-    expectation_suite_name: str = f"tmp.profiler_suite_{str(uuid.uuid4())[:8]}"
-    try:
-        # noinspection PyUnusedLocal
-        suite = context.get_expectation_suite(
-            expectation_suite_name=expectation_suite_name
-        )
-    except ge_exceptions.DataContextError:
-        suite = context.create_expectation_suite(
-            expectation_suite_name=expectation_suite_name
-        )
-        print(f'Created ExpectationSuite "{suite.expectation_suite_name}".')
-
     batch_request: dict = {
         "datasource_name": "taxi_pandas",
         "data_connector_name": "monthly",
         "data_asset_name": "my_reports",
     }
 
-    validator: Validator = context.get_validator(
-        batch_request=BatchRequest(**batch_request),
-        expectation_suite_name=expectation_suite_name,
+    validator: Validator = get_validator_with_temporary_expectation_suite(
+        component_name="profiler",
+        batch_request=batch_request,
+        data_context=context,
     )
     assert len(validator.batches) == 36
 
@@ -1870,23 +1805,7 @@ def test_quentin_expect_column_values_to_be_in_set_auto_yes_default_profiler_con
 ):
     context: DataContext = quentin_columnar_table_multi_batch_data_context
 
-    result: ExpectationValidationResult
-
     custom_profiler_config: RuleBasedProfilerConfig
-
-    suite: ExpectationSuite
-
-    expectation_suite_name: str = f"tmp.profiler_suite_{str(uuid.uuid4())[:8]}"
-    try:
-        # noinspection PyUnusedLocal
-        suite = context.get_expectation_suite(
-            expectation_suite_name=expectation_suite_name
-        )
-    except ge_exceptions.DataContextError:
-        suite = context.create_expectation_suite(
-            expectation_suite_name=expectation_suite_name
-        )
-        print(f'Created ExpectationSuite "{suite.expectation_suite_name}".')
 
     batch_request: dict = {
         "datasource_name": "taxi_pandas",
@@ -1894,14 +1813,15 @@ def test_quentin_expect_column_values_to_be_in_set_auto_yes_default_profiler_con
         "data_asset_name": "my_reports",
     }
 
-    validator: Validator = context.get_validator(
-        batch_request=BatchRequest(**batch_request),
-        expectation_suite_name=expectation_suite_name,
+    validator: Validator = get_validator_with_temporary_expectation_suite(
+        component_name="profiler",
+        batch_request=batch_request,
+        data_context=context,
     )
     assert len(validator.batches) == 36
 
     # Use all batches, loaded by Validator, for estimating Expectation argument values.
-    result = validator.expect_column_values_to_be_in_set(
+    result: ExpectationValidationResult = validator.expect_column_values_to_be_in_set(
         column="passenger_count",
         result_format="SUMMARY",
         include_config=True,
@@ -1941,23 +1861,7 @@ def test_quentin_expect_column_min_to_be_between_auto_yes_default_profiler_confi
 ):
     context: DataContext = quentin_columnar_table_multi_batch_data_context
 
-    result: ExpectationValidationResult
-
     custom_profiler_config: RuleBasedProfilerConfig
-
-    suite: ExpectationSuite
-
-    expectation_suite_name: str = f"tmp.profiler_suite_{str(uuid.uuid4())[:8]}"
-    try:
-        # noinspection PyUnusedLocal
-        suite = context.get_expectation_suite(
-            expectation_suite_name=expectation_suite_name
-        )
-    except ge_exceptions.DataContextError:
-        suite = context.create_expectation_suite(
-            expectation_suite_name=expectation_suite_name
-        )
-        print(f'Created ExpectationSuite "{suite.expectation_suite_name}".')
 
     batch_request: dict = {
         "datasource_name": "taxi_pandas",
@@ -1965,14 +1869,15 @@ def test_quentin_expect_column_min_to_be_between_auto_yes_default_profiler_confi
         "data_asset_name": "my_reports",
     }
 
-    validator: Validator = context.get_validator(
-        batch_request=BatchRequest(**batch_request),
-        expectation_suite_name=expectation_suite_name,
+    validator: Validator = get_validator_with_temporary_expectation_suite(
+        component_name="profiler",
+        batch_request=batch_request,
+        data_context=context,
     )
     assert len(validator.batches) == 36
 
     # Use all batches, loaded by Validator, for estimating Expectation argument values.
-    result = validator.expect_column_min_to_be_between(
+    result: ExpectationValidationResult = validator.expect_column_min_to_be_between(
         column="fare_amount",
         result_format="SUMMARY",
         include_config=True,
@@ -2011,36 +1916,21 @@ def test_quentin_expect_column_max_to_be_between_auto_yes_default_profiler_confi
 ):
     context: DataContext = quentin_columnar_table_multi_batch_data_context
 
-    result: ExpectationValidationResult
-
-    suite: ExpectationSuite
-
-    expectation_suite_name: str = f"tmp.profiler_suite_{str(uuid.uuid4())[:8]}"
-    try:
-        # noinspection PyUnusedLocal
-        suite = context.get_expectation_suite(
-            expectation_suite_name=expectation_suite_name
-        )
-    except ge_exceptions.DataContextError:
-        suite = context.create_expectation_suite(
-            expectation_suite_name=expectation_suite_name
-        )
-        print(f'Created ExpectationSuite "{suite.expectation_suite_name}".')
-
     batch_request: dict = {
         "datasource_name": "taxi_pandas",
         "data_connector_name": "monthly",
         "data_asset_name": "my_reports",
     }
 
-    validator: Validator = context.get_validator(
-        batch_request=BatchRequest(**batch_request),
-        expectation_suite_name=expectation_suite_name,
+    validator: Validator = get_validator_with_temporary_expectation_suite(
+        component_name="profiler",
+        batch_request=batch_request,
+        data_context=context,
     )
     assert len(validator.batches) == 36
 
     # Use all batches, loaded by Validator, for estimating Expectation argument values.
-    result = validator.expect_column_max_to_be_between(
+    result: ExpectationValidationResult = validator.expect_column_max_to_be_between(
         column="fare_amount",
         result_format="SUMMARY",
         include_config=True,
@@ -2107,29 +1997,16 @@ def test_quentin_expect_column_unique_value_count_to_be_between_auto_yes_default
 
     result: ExpectationValidationResult
 
-    suite: ExpectationSuite
-
-    expectation_suite_name: str = f"tmp.profiler_suite_{str(uuid.uuid4())[:8]}"
-    try:
-        # noinspection PyUnusedLocal
-        suite = context.get_expectation_suite(
-            expectation_suite_name=expectation_suite_name
-        )
-    except ge_exceptions.DataContextError:
-        suite = context.create_expectation_suite(
-            expectation_suite_name=expectation_suite_name
-        )
-        print(f'Created ExpectationSuite "{suite.expectation_suite_name}".')
-
     batch_request: dict = {
         "datasource_name": "taxi_pandas",
         "data_connector_name": "monthly",
         "data_asset_name": "my_reports",
     }
 
-    validator: Validator = context.get_validator(
-        batch_request=BatchRequest(**batch_request),
-        expectation_suite_name=expectation_suite_name,
+    validator: Validator = get_validator_with_temporary_expectation_suite(
+        component_name="profiler",
+        batch_request=batch_request,
+        data_context=context,
     )
     assert len(validator.batches) == 36
