+++ tests/rule_based_profiler/data_assistant/test_volume_data_assistant.py
@@ -1,19 +1,10 @@
-import uuid
-from typing import Any, Dict, List, Optional, Union
+from typing import Any, Dict, List
 from unittest import mock
 
 from freezegun import freeze_time
 
-import great_expectations.exceptions as ge_exceptions
 from great_expectations import DataContext
 from great_expectations.core import ExpectationConfiguration, ExpectationSuite
-from great_expectations.core.batch import (
-    BatchRequest,
-    BatchRequestBase,
-    RuntimeBatchRequest,
-    materialize_batch_request,
-)
-from great_expectations.data_context import BaseDataContext
 from great_expectations.rule_based_profiler.config import RuleBasedProfilerConfig
 from great_expectations.rule_based_profiler.data_assistant import (
     DataAssistant,
@@ -36,6 +27,7 @@ from great_expectations.rule_based_profiler.types.data_assistant_result import (
 from great_expectations.util import deep_filter_properties_iterable
 from great_expectations.validator.validator import Validator
 from tests.rule_based_profiler.parameter_builder.conftest import RANDOM_SEED
+from tests.test_utils import get_validator_with_temporary_expectation_suite
 
 
 # TODO: <Alex>ALEX -- Make this helper utility of general use.</Alex>
@@ -59,33 +51,6 @@ def set_bootstrap_random_seed_variable(
         )
 
 
-# TODO: <Alex>ALEX -- Make this helper utility of general use.</Alex>
-def get_validator_with_temporary_expectation_suite(
-    data_context: BaseDataContext, batch_request: Union[BatchRequestBase, dict]
-) -> Validator:
-    suite: ExpectationSuite
-
-    expectation_suite_name: str = f"tmp.profiler_suite_{str(uuid.uuid4())[:8]}"
-    try:
-        # noinspection PyUnusedLocal
-        suite = data_context.get_expectation_suite(
-            expectation_suite_name=expectation_suite_name
-        )
-    except ge_exceptions.DataContextError:
-        suite = data_context.create_expectation_suite(
-            expectation_suite_name=expectation_suite_name
-        )
-        print(f'Created ExpectationSuite "{suite.expectation_suite_name}".')
-
-    batch_request = materialize_batch_request(batch_request=batch_request)
-    validator: Validator = data_context.get_validator(
-        batch_request=batch_request,
-        expectation_suite_name=expectation_suite_name,
-    )
-
-    return validator
-
-
 @freeze_time("09/26/2019 13:42:41")
 def test_get_metrics_and_expectations(
     quentin_columnar_table_multi_batch_data_context,
@@ -99,8 +64,9 @@ def test_get_metrics_and_expectations(
     }
 
     validator: Validator = get_validator_with_temporary_expectation_suite(
-        data_context=context,
+        component_name="data_assistant",
         batch_request=batch_request,
+        data_context=context,
     )
     assert len(validator.batches) == 36
 
@@ -2566,8 +2532,7 @@ def test_get_metrics_and_expectations(
     # Utilize a consistent seed to deal with probabilistic nature of this feature.
     data_assistant: DataAssistant = VolumeDataAssistant(
         name="test_volume_data_assistant",
-        batch_request=batch_request,
-        data_context=context,
+        validator=validator,
     )
     set_bootstrap_random_seed_variable(profiler=data_assistant.profiler)
     data_assistant_result: DataAssistantResult = data_assistant.run(
@@ -2607,8 +2572,9 @@ def test_execution_time_within_proper_bounds(
     }
 
     validator: Validator = get_validator_with_temporary_expectation_suite(
-        data_context=context,
+        component_name="data_assistant",
         batch_request=batch_request,
+        data_context=context,
     )
     assert len(validator.batches) == 36
 
@@ -2634,8 +2600,9 @@ def test_volume_data_assistant_plot_descriptive(
     }
 
     validator: Validator = get_validator_with_temporary_expectation_suite(
-        data_context=context,
+        component_name="data_assistant",
         batch_request=batch_request,
+        data_context=context,
     )
     assert len(validator.batches) == 36
 
@@ -2668,8 +2635,9 @@ def test_volume_data_assistant_plot_prescriptive(
     }
 
     validator: Validator = get_validator_with_temporary_expectation_suite(
-        data_context=context,
+        component_name="data_assistant",
         batch_request=batch_request,
+        data_context=context,
     )
     assert len(validator.batches) == 36
